logs/requirements.txt
# REST API Generation
fastapi==0.85.0
uvicorn==0.18.2
pydantic==1.9.1
SQLAlchemy==1.4.39
aiosqlite==0.17.0
coloredlogs==15.0.1
PyYAML==6.0
requests==2.31.0
aio-pika==9.3.0
asyncio==3.4.3
PyJWT==2.8.0
cryptography==41.0.5
python-consul2==0.1.5
python-dotenv==0.21.0
dnspython==2.2.1
ifaddr==0.2.0